+++ src/modules/layerInformation/components/LayerInformation.vue
@@ -112,9 +112,9 @@ export default {
             if (layer.typ === "GROUP" && layer.layers) {
                 const selectedLayer = layer.layers[this.selectedOption];
 
-                return selectedLayer ? this.getGetCapabilitiesUrl(selectedLayer) : "";
+                return selectedLayer ? this.getLayerAddress(selectedLayer) : "";
             }
-            return layer.url ? this.getGetCapabilitiesUrl(layer) : "";
+            return layer.url ? this.getLayerAddress(layer) : "";
         },
         legendURL  () {
             return this.layerInfo.legendURL;
@@ -273,19 +273,45 @@ export default {
             return {active: this.isActiveTab(tab), show: this.isActiveTab(tab), "tab-pane": true, fade: true};
         },
         /**
-         * generates a GetCapabilities URL from a given service base address and type
+         * Removes slash or questionmark from the end of the url.
+         * @param {String} url the url to clean
+         * @returns {String} the cleaned URL
+         */
+        cleanUrl (url) {
+            let baseUrl = url.split("?")[0];
+
+            if (baseUrl.endsWith("/")) {
+                baseUrl = baseUrl.slice(0, -1);
+            }
+            return baseUrl;
+        },
+        /**
+         * Generates a GetCapabilities URL from a given service base address and type vor 2D layers or
+         * Appends 'tileset.json' to TileSet3D layers url, if not exists.
+         * Appends always 'layer.json' to Terrain3D layers url.
          * @param {Object} param payload
          * @param {String} param.url service base URL
          * @param {String} param.typ service type (e.g., WMS)
-         * @returns {String} GetCapabilities URL
+         * @returns {String} the created URL
          */
-        getGetCapabilitiesUrl (layerInfo) {
+        getLayerAddress (layerInfo) {
             const typ = layerInfo.typ,
-                config = this.layerConfigById(layerInfo.id),
-                url = config?.origUrl ? config.origUrl : layerInfo.url,
+                config = this.layerConfigById(layerInfo.id);
+            let url = config?.origUrl ? config.origUrl : layerInfo.url,
                 urlObject = new URL(url, location.href);
 
-            if (typ !== "OAF") {
+            if (typ.toUpperCase() === "TILESET3D") {
+                const baseUrl = this.cleanUrl(url);
+
+                // if no json-file is provided in masterportalAPI src/layer/tileset.js "/tileset.json" is appended
+                url = baseUrl + (baseUrl.endsWith(".json") ? "" : "/tileset.json");
+                urlObject = new URL(url, location.href);
+            }
+            else if (typ.toUpperCase() === "TERRAIN3D") {
+                // terrain layer: in Cesium "/layer.json" is appended to url
+                urlObject = new URL(this.cleanUrl(url) + "/layer.json", location.href);
+            }
+            else if (typ !== "OAF") {
                 urlObject.searchParams.set("SERVICE", typ);
                 urlObject.searchParams.set("REQUEST", "GetCapabilities");
             }